refactor assets screen
